requirements.txt
transformers
torch
requests
pillow
fastapi
uvicorn
streamlit
python-multipart
llama-index
pypdf
openai
python-dotenv
vectara
ipykernel
altair
blinker
cachetools
click
gitpython
numpy
packaging
pandas
pillow
protobuf
pyarrow
pydeck
requests
rich
tenacity
toml
tornado
typing-extensions
watchdog
streamlit
typing
pprint
Bio
llama-index